backend/requirements.txt
flask>=2.2
flask-cors>=3.0.10
torch>=2.0
torchvision
opencv-python-headless
pillow
numpy 